+++ lib/rubocop/cop/performance/count.rb
@@ -25,6 +25,19 @@ module RuboCop
       #   [1, 2, 3].count { |e| e < 2 && e.even? }
       #   Model.select('field AS field_one').count
       #   Model.select(:value).count
+      #
+      # `ActiveRecord` compatibility:
+      # `ActiveRecord` will ignore the block that is passed to `count`.
+      # Other methods, such as `select`, will convert the association to an
+      # array and then run the block on the array. A simple work around to
+      # make `count` work with a block is to call `to_a.count {...}`.
+      #
+      # Example:
+      #   Model.where(id: [1, 2, 3].select { |m| m.method == true }.size
+      #
+      #   becomes:
+      #
+      #   Model.where(id: [1, 2, 3]).to_a.count { |m| m.method == true }
       class Count < Cop
         MSG = 'Use `count` instead of `%s...%s`.'.freeze
 
@@ -32,6 +45,7 @@ module RuboCop
         COUNTERS = [:count, :length, :size].freeze
 
         def on_send(node)
+          return unless should_run?
           selector, selector_loc, params, counter = parse(node)
           return unless COUNTERS.include?(counter)
           return unless SELECTORS.include?(selector)
@@ -61,6 +75,12 @@ module RuboCop
         end
 
         private
+
+        def should_run?
+          !(cop_config['SafeMode'.freeze] ||
+            config['Rails'.freeze] &&
+            config['Rails'.freeze]['Enabled'.freeze])
+        end
 
         def parse(node)
           left, counter = *node